+++ suite-common/wallet-core/src/fiat-rates/fiatRatesConstants.ts
